Moves some test only code out of legacy_implementation/db.py
